fix: add missing admin prefix to category routes

+++ routes/web.php
@@ -20,13 +20,13 @@ Route::middleware('auth')->group(function () {
     Route::patch('/profile', [ProfileController::class, 'update'])->name('profile.update');
     Route::delete('/profile', [ProfileController::class, 'destroy'])->name('profile.destroy');
 
-    Route::get('/categories', [CategoryController::class, 'index'])->name('admin.categories.index');
-    Route::get('/categories/create', [CategoryController::class, 'create'])->name('admin.categories.create');
-    Route::get('/categories/{category}', [CategoryController::class, 'show'])->name('admin.categories.show');
-    Route::get('/categories/{category}/edit', [CategoryController::class, 'edit'])->name('admin.categories.edit');
-    Route::post('/categories', [CategoryController::class, 'store'])->name('admin.categories.store');
-    Route::put('/categories/{category}', [CategoryController::class, 'update'])->name('admin.categories.update');
-    Route::delete('/categories/{category}', [CategoryController::class, 'destroy'])->name('admin.categories.destroy');
+    Route::get('admin/categories', [CategoryController::class, 'index'])->name('admin.categories.index');
+    Route::get('admin/categories/create', [CategoryController::class, 'create'])->name('admin.categories.create');
+    Route::get('admin/categories/{category}', [CategoryController::class, 'show'])->name('admin.categories.show');
+    Route::get('admin/categories/{category}/edit', [CategoryController::class, 'edit'])->name('admin.categories.edit');
+    Route::post('admin/categories', [CategoryController::class, 'store'])->name('admin.categories.store');
+    Route::put('admin/categories/{category}', [CategoryController::class, 'update'])->name('admin.categories.update');
+    Route::delete('admin/categories/{category}', [CategoryController::class, 'destroy'])->name('admin.categories.destroy');
 
     Route::get('/admin/products', [ProductController::class, 'index'])->name('admin.products.index');
     Route::get('/admin/products/create', [ProductController::class, 'create'])->name('admin.products.create');
